add hook to keep canvas logic in
